+++ oracle-simple.cabal
@@ -33,7 +33,7 @@ library
   exposed-modules:
     Database.Oracle.Simple
   other-modules:
-    Database.Oracle.Simple.Internal
+    Database.Oracle.Simple.Internal, Database.Oracle.Simple.FromRow, Database.Oracle.Simple.FromField
   build-depends:
     base < 5, derive-storable, text
   hs-source-dirs:
