+++ src/Database/Oracle/Simple.hs
@@ -1,3 +1,5 @@
 module Database.Oracle.Simple (module X) where
 
+import Database.Oracle.Simple.FromField as X
 import Database.Oracle.Simple.Internal as X
+import Database.Oracle.Simple.FromRow as X
